[LF-66/elasticsearch-keyword] 프로젝트 패키지 구조 변경
