+++ LiveFeedService/src/main/java/com/livefeed/livefeedservice/articlelist/dto/ArticleListDto.java
@@ -10,9 +10,15 @@ public record ArticleListDto(
         String pit
 ) {
 
-    public static ArticleListDto of(List<ArticleDto> articles, boolean isLast, Float lastScore, Long lastId, String pit) {
+    public static ArticleListDto from(SearchResultDto searchResultDto, int pageSize) {
+        List<ArticleDto> articleList = searchResultDto.articleDtoList();
+
         return new ArticleListDto(
-                articles, isLast, lastScore, lastId, pit
+                articleList,
+                articleList.size() < pageSize,
+                articleList.size() > 0 ? articleList.get(articleList.size() - 1).score() : null,
+                articleList.size() > 0 ? articleList.get(articleList.size() - 1).articleId() : null,
+                searchResultDto.pit()
         );
     }
 }
