+++ LiveFeedService/src/main/java/com/livefeed/livefeedservice/articlelist/dto/SearchResultDto.java
@@ -1,0 +1,9 @@
+package com.livefeed.livefeedservice.articlelist.dto;
+
+import java.util.List;
+
+public record SearchResultDto(
+        List<ArticleDto> articleDtoList,
+        String pit
+) {
+}
