+++ LiveFeedService/src/main/java/com/livefeed/livefeedservice/articlelist/repository/SearchOperations.java
@@ -1,5 +1,6 @@
 package com.livefeed.livefeedservice.articlelist.repository;
 
+import com.livefeed.livefeedservice.articlelist.dto.SearchResultDto;
 import com.livefeed.livefeedservice.articlelist.util.SearchQueryParam;
 import com.livefeed.livefeedservice.elasticsearch.entity.ElasticsearchArticle;
 import org.springframework.data.elasticsearch.core.SearchHits;
@@ -8,5 +9,5 @@ public interface SearchOperations {
 
     String getPit(String indexName);
 
-    SearchHits<ElasticsearchArticle> getSearchResult(SearchQueryParam searchQueryParam);
+    SearchResultDto getSearchResultTemp(SearchQueryParam searchQueryParam);
 }
