+++ LiveFeedService/src/test/java/com/livefeed/livefeedservice/articlelist/controller/ArticleListControllerTest.java
@@ -2,6 +2,7 @@ package com.livefeed.livefeedservice.articlelist.controller;
 
 import com.livefeed.livefeedservice.articlelist.dto.ArticleDto;
 import com.livefeed.livefeedservice.articlelist.dto.ArticleListDto;
+import com.livefeed.livefeedservice.articlelist.dto.SearchResultDto;
 import com.livefeed.livefeedservice.articlelist.service.ArticleListService;
 import com.livefeed.livefeedservice.articlelist.util.SearchQueryParam;
 import org.junit.jupiter.api.BeforeEach;
@@ -53,16 +54,12 @@ class ArticleListControllerTest {
     @DisplayName("search 쿼리를 통해 기사를 검색할 수 있습니다.")
     void getArticleList() throws Exception {
         // given
-        ArticleListDto data = ArticleListDto.of(
-                List.of(
-                        new ArticleDto(5L, "article5", "press5", "content5", "photo5", "5분전", 0f),
-                        new ArticleDto(3L, "article3", "press3", "content3", "photo3", "3분전", 0f)
-                ),
-                false,
-                0f,
-                3L,
-                "3eaGBAEIYXJ0aWNsZXMWOVpGTDhfYTh"
-        );
+        SearchResultDto searchResultDto = new SearchResultDto(List.of(
+                new ArticleDto(5L, "article5", "press5", "content5", "photo5", "5분전", 0f),
+                new ArticleDto(3L, "article3", "press3", "content3", "photo3", "3분전", 0f)
+        ), "3eaGBAEIYXJ0aWNsZXMWOVpGTDhfYTh");
+
+        ArticleListDto data = ArticleListDto.from(searchResultDto, 2);
 
         Mockito.doReturn(data).when(articleListService).getArticleList(Mockito.any(SearchQueryParam.class));
 
@@ -99,35 +96,5 @@ class ArticleListControllerTest {
                 .andExpect(MockMvcResultMatchers.jsonPath("data.isLast").value(false))
                 .andExpect(MockMvcResultMatchers.jsonPath("data.lastId").value(3L))
                 .andExpect(MockMvcResultMatchers.jsonPath("data.pit").value("3eaGBAEIYXJ0aWNsZXMWOVpGTDhfYTh"));
-//                .andDo(
-//                        MockMvcRestDocumentation.document("article-list",
-//                                Preprocessors.preprocessRequest(Preprocessors.prettyPrint()),
-//                                Preprocessors.preprocessResponse(Preprocessors.prettyPrint()),
-//                                RequestDocumentation.queryParameters(
-//                                        RequestDocumentation.parameterWithName("type").description("제목/본문 어디서 검색할지에 대한 조건").optional(),
-//                                        RequestDocumentation.parameterWithName("keyword").description("검색하고자 하는 키워드").optional(),
-//                                        RequestDocumentation.parameterWithName("size").description("한번에 검색하는 목록 개수").optional(),
-//                                        RequestDocumentation.parameterWithName("sort").description("정렬 기준 (id-desc)").optional(),
-//                                        RequestDocumentation.parameterWithName("lastId").description("마지막으로 조회된 기사 id 값").optional(),
-//                                        RequestDocumentation.parameterWithName("pit").description("기존 검색에서 전달받은 pit 값").optional()
-//                                ),
-//                                PayloadDocumentation.responseFields(
-//                                        PayloadDocumentation.fieldWithPath("success").type(JsonFieldType.BOOLEAN).description("성공 여부"),
-//                                        PayloadDocumentation.fieldWithPath("status").type(JsonFieldType.NUMBER).description("상태값"),
-//                                        PayloadDocumentation.fieldWithPath("message").type(JsonFieldType.STRING).description("상세 메시지"),
-//                                        PayloadDocumentation.fieldWithPath("data").type(JsonFieldType.OBJECT).description("응답 데이터"),
-//                                        PayloadDocumentation.fieldWithPath("data.articles[0].articleId").type(JsonFieldType.NUMBER).description("기사 아이디"),
-//                                        PayloadDocumentation.fieldWithPath("data.articles[0].title").type(JsonFieldType.STRING).description("기사 제목"),
-//                                        PayloadDocumentation.fieldWithPath("data.articles[0].pressCompany").type(JsonFieldType.STRING).description("언론사 이름"),
-//                                        PayloadDocumentation.fieldWithPath("data.articles[0].content").type(JsonFieldType.STRING).description("기사 본문"),
-//                                        PayloadDocumentation.fieldWithPath("data.articles[0].photo").type(JsonFieldType.STRING).description("기사의 첫 번째 사진"),
-//                                        PayloadDocumentation.fieldWithPath("data.articles[0].minutesAgo").type(JsonFieldType.STRING).description("기사의 발행 시간"),
-//                                        PayloadDocumentation.fieldWithPath("data.isLast").type(JsonFieldType.BOOLEAN).description("마지막 페이지인지 여부"),
-//                                        PayloadDocumentation.fieldWithPath("data.lastId").type(JsonFieldType.NUMBER).description("응답 기사 목록 중 마지막 기사의 id"),
-//                                        PayloadDocumentation.fieldWithPath("data.pit").type(JsonFieldType.STRING).description("해당 검색에 해당하는 point in time 값")
-//
-//                                )
-//                        )
-//                );
     }
 }
