+++ http/LiveFeedService/articledetail/articledetail.http
@@ -2,8 +2,8 @@
 GET http://localhost:8084/api/detail/articles/1
 
 ### 기사 목록을 전달하는 api
-GET http://localhost:8084/api/list/articles?type=headerHtml&keyword=손흥민&size=3&sort=id-desc
+GET http://localhost:8084/api/list/articles?type=articleTitle,bodyHtml&keyword=바르셀로나,김민재&size=10&related=true
 #GET http://localhost:8084/api/list/articles?type=articleTitle&keyword=페이커&size=4&sort=id-desc
 
 ### search 쿼리가 아닌 api
-GET http://localhost:8084/api/list/articles?size=4&sort=id-desc
+GET http://localhost:8084/api/list/articles?size=4&related=true
